+++ lib/Braintree/LocalPaymentCompleted.php
@@ -17,11 +17,10 @@ namespace Braintree;
  * @package    Braintree
  * @category   Resources
  *
- * @property-read string $grantOwnerMerchantId
- * @property-read string $grantRecipientMerchantId
+ * @property-read string $paymentId
+ * @property-read string $payerId
  * @property-read string $paymentMethodNonce
- * @property-read string $token
- * @property-read string $updatedFields
+ * @property-read \Braintree\Transaction $transaction
  */
 class LocalPaymentCompleted extends Base
 {
@@ -52,6 +51,14 @@ class LocalPaymentCompleted extends Base
     {
         // set the attributes
         $this->_attributes = $localPaymentCompletedAttribs;
+
+        if (isset($transactionAttribs['transaction'])) {
+            $this->_set('transaction',
+                new Transaction(
+                    $transactionAttribs['transaction']
+                )
+            );
+        }
     }
 
     /**